more mps settings

+++ utils/general.py
@@ -263,7 +263,7 @@ def labels_to_class_weights(labels, nc=80):
     weights[weights == 0] = 1  # replace empty bins with 1
     weights = 1 / weights  # number of targets per class
     weights /= weights.sum()  # normalize
-    return torch.from_numpy(weights)
+    return torch.from_numpy(weights.astype(np.float32))
 
 
 def labels_to_image_weights(labels, nc=80, class_weights=np.ones(80)):
